Amendments to unit tests to remove use of build_coefficients_cubelist helper function and removal of unnecessary checks in create_coefficients_cubelist method and simplication of metadata.
